add middleware for check perms

+++ internal/modules/event/entity/event.go
@@ -8,7 +8,7 @@ import (
 
 type Event struct {
 	gorm.Model
-	ID             uint64 `gorm:"primaryKey;autoIncrement"`
+	ID             uint `gorm:"primaryKey;autoIncrement"`
 	Regular        bool
 	DateEnd        string
 	CreatedAt      time.Time
